Added tests to cover new feature:: getting a number from a string representing roman numerals. Code will not yet compile.
The method should make use of NullReference and Argument exceptions when dealing with null/empty strings respectively. The method should also throw ArgumentException when strings contain non-numeral characters or when lowercase symbol characters are used. Subtractive notation should be enforced and throw ArgumentException on violations. Finally, the same number range between 1 and 3000 should be enforced on the return values and throw IndexOutOfRangeException on violations.